Refactor file paths to align with new extension structure and update imports accordingly
